requirements.txt
# ─── Discord ───────────────────────────────────────────────────────────────────
discord.py==2.3.2

# ─── Telegram ──────────────────────────────────────────────────────────────────
aiogram==3.7.0

# ─── Database ──────────────────────────────────────────────────────────────────
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1

# ─── Cache / Queue ─────────────────────────────────────────────────────────────
redis==5.0.4
aioredis==2.0.1

# ─── Config / Validation ───────────────────────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.3.1
python-dotenv==1.0.1

# ─── Logging ───────────────────────────────────────────────────────────────────
structlog==24.2.0

# ─── Utilities ─────────────────────────────────────────────────────────────────
python-dateutil==2.9.0
humanize==4.9.0